Add working like clicks
