+++ src/index.html
@@ -9,12 +9,13 @@
     <script src='https://kit.fontawesome.com/a076d05399.js' crossorigin='anonymous'></script>
 </head>
 <body>
-    <body class="container text-center">
+    <div class="container text-center">
         <h1>Search TV MAZE</h1>
         <div class="row justify-content-start gap-4 shows">
         
         </div>
         <span class="text-danger errorMessage"></span>
-    </body>
+        <ul class="like-value"></ul>
+    </div>
 </body>
 </html>
