+++ src/styles/style.css
@@ -14,6 +14,6 @@ p {
   text-overflow: ellipsis;
 }
 
-.like {
+.like-button {
   cursor: pointer;
 }
